+++ leia-schema-validator/src/test/java/com/grookage/leia/validator/stubs/TestEnum.java
@@ -1,0 +1,6 @@
+package com.grookage.leia.validator.stubs;
+
+public enum TestEnum {
+    ONE,
+    TWO;
+}